requirements.txt
moviepy>=2.0
Pillow>=10.0
numpy>=1.26
soundfile>=0.12
librosa>=0.10
opencv-python>=4.8
pygame>=2.5
